add useSpecimensOptions hook
