+++ themes/gatsby-theme-specimens/index.js
@@ -3,5 +3,6 @@ import ColorSwatch from "./src/components/color-swatch"
 import ColorRow from "./src/components/color-row"
 import Palette from "./src/components/palette"
 import normalizeThemeUI from "./src/utils/normalize-theme-ui"
+import filterStyles from "./src/utils/filter-styles"
 
-export { Alert, ColorSwatch, ColorRow, Palette, normalizeThemeUI }
+export { Alert, ColorSwatch, ColorRow, Palette, normalizeThemeUI, filterStyles }
