+++ themes/gatsby-theme-specimens/src/utils/filter-styles.ts
@@ -1,0 +1,12 @@
+const filterStyles = ({ input, allowed }) =>
+  Object.keys(input)
+    .filter(key => allowed.includes(key))
+    .reduce(
+      (obj, key) => ({
+        ...obj,
+        [key]: input[key],
+      }),
+      {}
+    )
+
+export default filterStyles
